fix: moving package versions to Directory.Packages.props from csproj in init command
When nunit tests project is created with dotnet new then nugets with versions are added to csproj (and restored) instead of Directory.Packages.props (this file is created later in the process, but even if it's present, nugets are put fully in csproj - it's probably dotnet cli bug). So I need to move versions to Directory.Packages.props.