build: changed test reporter to artifact scenario
This is because previous implementation could place test results under another workflow. Test report action is using check run functionality, which is working for given commit, not workflow. So if we have more than 1 workflow for 1 commit, then results can be placed in a random workflow. With current way, it seems to work fine.