build: moved format workflow to build workflow as a job
This is to have 1 workflow as a workaround for GitHub Checks API weird behaviour. Test reporter action is using this API to publish test results. Checks API result is attached to commit, not workflow, so result can appear in a random workflow for given commit. To avoid that it is better to have 1 workflow with 2 jobs.